requirements.txt
# Robot Framework Core
robotframework>=6.1.0
robotframework-requests>=0.9.4
robotframework-faker>=6.0.0
robotframework-allure>=2.13.2

# HTTP Client
requests>=2.31.0

# Data handling
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Messaging
pika>=1.3.2
kafka-python>=2.0.2

# Monitoring & Metrics
prometheus-client>=0.17.1
psutil>=5.9.5

# Testing utilities
pytest>=7.4.0
pytest-html>=3.2.0

# Utilities
jsonpath-ng>=1.5.3
Faker>=19.6.0
